pyproject.toml
[tool.poetry]
name = "clothing-store"
version = "0.1.0"
description = ""
authors = ["Arybjan <arybjen92@gmail.com>"]
readme = "README.md"
packages = [{include = "clothing_store"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
